MainWindow is created and ChartView is used as a widget on it
A simple MainWindow is created and ChartView is promoted as
a GraphicsView on the central widget of main window.
The window created in the main function is changed to
the new MainWindow. The k_means object creation and
clustering is moved into the MainWindow constructor.
paintPoints function in the ChartView is used to
draw points in the chart as a scatter plot.
Currently, only the all points are plotted in the
chart, nothing on the clustering is plotted in it.