Initialization class is created to provide multiple options
Initialization class is created to provide multiple options
of center initialization for the clustering. Init random
sample function is moved to the created initialization class
from k_means class. Getters for the class attributes in k_means
are implemented to be able to do the initialization from the new
class. addCluster function is added to k_means to be used
in initialization class. A fix is done in the promotion of the
ChartView in MainWindow.ui