Functionalities of generate and cluster points are implemented
Functionalities of Generate points and cluster points buttons
in the dialog are implemented. Necessary attributes and functions
are added to mainwindow and k_means class. FIX TODO: Currently
the number of clusters is k+1 since indexing start from zero,
needs to be fixed so that we have exactly k clusters