Initialization, point size, clustering by steps are added to ui
Initializing the clustering is added to the dialog as a separate
combobox selection for user with corresponding initalize button.
Initialize button is connected and its functionality is tested.
Clustering with steps is added as another button, connections are
made, not tested yet. For visualization, point size is also
added as a parameter that user can define. No connections are done
for that functionality yet.