Updating point sizes is allowed except after clustering started
Point size can be entered and the point size in the chart can be
updated with new created spinbox and corresponding update chart
button. Currently it only works when the points are generated,
and when clustering is initialized. Once clustering starts,
user cannot change the point size. TODO: Will fix if time.
Extra functions created in attempt of panning are deleted.