New design class called ScatterWidget3D is created
New design class called ScatterWidget3D is created.
This class will use Q3DScatter class for 3D point drawing
in the 3D version of the kmeans project. Another scroll
area is added to the MainWindow and an instance of
ScatterWidget3D is added inside that scroll area.
Widgets are placed in a way that 2D is at the upper half
of the mainwindow, whereas 3D is at lower half of the
mainwindow. Will try to make them visible in two separate
boxes if time.