Importing points from a txt file functionality is implemented
A button to import points from a txt file is added to the
kmeans dialog. Also, one label is used at the bottom of the
two buttons allowing user to generate or import points, to
inform user whether the points were generated/imported.
Imported points are set in the kmeans object inside
mainwindow. Currently only tested with 2D points.