Update point size feature is implemented for 3D visualization
Allowing user to update the point size in the 3D visualization
is implemented for 3D. Size 0 means the graph will automatically
resize the points with respect to the number of points. Otherwise
the size is determined by the spinbox value from ui / 100.0
because the item size is between 0.0f and 1.0f in 3D vis.