Checks for special cases for K and number of points are added
Checks for special cases for K and number of points are added
to initialization and generation functions in mainwindow.
When K = 0, K > numOfPoints, N = 0 are not allowed, an error
message pops up when these cases occur so that user can
fix the values to avoid these cases. K = 1 is tested for
both 2D and 3D.