Manual selection of cluster centers in 3D is implemented
Manual selection of cluster centers in 3D is implemented.
A mouse press event for left mouse button is overriden
in Scatter3DWidget. This event takes the selected point
from the 3DScatter and calls the selectClusterCenter function
which triggers a signal connected to corresponding slot
in the MainWindow. Once a point is selected manually,
the initialization parameters are collected from dialog ui,
cluster centers are added as user selects points and
clicks to the Scatter3DWidget on MainWindow to add them
as clusters. Each added cluster center is visualized on the
3DScatter. Once all centers are selected, an error is popped
up if user tries to select more. When all are selected,
clustering can be run using corresponding buttons.