Panning in 3D is implemented allowing user to move in x z axis
Panning in 3D is implemented by implementing a Q3DInputHandler
class named InputHandlerPanning. This implementation is found
and modified from the QT examples. With this feature, the user
can move along x and z axis by clicking on the identifier
values at the axis and move in the direction of that axis.