Extra error check for the left click on Scatter3DWidget is added
Extra error check for the left click on Scatter3DWidget is added.
This overriden mouse event is used to manually initialize 3D
cluster centers. There was no check to make sure the points are
already generated or imported. Therefore, the program was
crashing when someone clicks on that area before any data is
generated or imported. After the fix, the program does not crash.